Move org services and controllers into their own folder structure
